requirements.txt
python3-netsnmp
pytest~=7.2.0
